Updates for changed IC2 API. Not sure if I'll be able to use @Optional with the new non-legacy BC MJ API as it uses Annotations instead of interfaces.

+++ src/main/java/mekanism/common/item/ItemBlockEnergyCube.java
@@ -142,7 +142,7 @@ public class ItemBlockEnergyCube extends ItemBlock implements IEnergizedItem, IE
 
 	@Override
 	@Method(modid = "IC2API")
-	public int getMaxCharge(ItemStack itemStack)
+	public double getMaxCharge(ItemStack itemStack)
 	{
 		return 0;
 	}
@@ -156,7 +156,7 @@ public class ItemBlockEnergyCube extends ItemBlock implements IEnergizedItem, IE
 
 	@Override
 	@Method(modid = "IC2API")
-	public int getTransferLimit(ItemStack itemStack)
+	public double getTransferLimit(ItemStack itemStack)
 	{
 		return 0;
 	}
