+++ src/main/java/mekanism/common/item/ItemEnergized.java
@@ -87,7 +87,7 @@ public class ItemEnergized extends ItemMekanism implements IEnergizedItem, ISpec
 
 	@Override
 	@Method(modid = "IC2API")
-	public int getMaxCharge(ItemStack itemStack)
+	public double getMaxCharge(ItemStack itemStack)
 	{
 		return 0;
 	}
@@ -101,7 +101,7 @@ public class ItemEnergized extends ItemMekanism implements IEnergizedItem, ISpec
 
 	@Override
 	@Method(modid = "IC2API")
-	public int getTransferLimit(ItemStack itemStack)
+	public double getTransferLimit(ItemStack itemStack)
 	{
 		return 0;
 	}
